+++ apps/admin/src/main.tsx
@@ -1,13 +1,19 @@
 import { StrictMode } from 'react';
-import * as ReactDOM from 'react-dom/client';
+import * as ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
-import App from './app/app';
+import { App } from './app/app';
+import { fetchConfiguration } from './config';
+import { getConfig } from './environments/environment';
 
-const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement,
-);
-root.render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-);
+fetchConfiguration({ getConfig })
+  .then((config) => {
+    const container = document.getElementById('root') as Element;
+    const root = createRoot(container);
+    root.render(
+      <StrictMode>
+        <App config={config} />
+      </StrictMode>,
+    );
+  })
+  .catch((err) => console.error(err));