+++ apps/docs/deploy.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+
+import { deploy } from '@docusaurus/core/lib/commands/deploy';
+
+deploy(__dirname, {
+  skipBuild: true,
+  outDir: '../../dist/apps/docs',
+}).catch((e) => {
+  console.error(e);
+  process.exit(1);
+});